exception handler added, controller implements interface
