+++ conveyor/src/main/java/com/example/conveyor/exception/ConveyorApiExceptionHandler.java
@@ -1,0 +1,17 @@
+package com.example.conveyor.exception;
+
+import org.springdoc.api.ErrorMessage;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ConveyorApiExceptionHandler {
+    @ExceptionHandler(ScoringException.class)
+    public ResponseEntity<ErrorMessage> handleException(ScoringException exception) {
+        return ResponseEntity
+                .status(HttpStatus.BAD_REQUEST)
+                .body(new ErrorMessage(exception.getMessage()));
+    }
+}
